test(config): move TestUpdateConfig to config_test.go
